Create helper class for building joins
A large chunk of the code in the DataflowToSqlQueryPlanConverter
and the module contianing it is dedicated to managing the various
mechanics of constructing joins. Some of this is overlapping in
whole or in parts, but refactoring the join-specific operations
is challenging because of how spread out everything is within this
file.

This is step one in moving the join building logic into a more
focused place - creating a helper class and migrating one
static function into it. Future commits will move other pieces
around, and then we can evaluate the extent to which we can
consolidate and re-use some of this logic.
