+++ metricflow/plan_conversion/dataflow_to_sql.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import logging
 from collections import OrderedDict
-from dataclasses import dataclass
 from typing import Generic, List, Optional, Sequence, Tuple, TypeVar, Union
 
 from metricflow.aggregation_properties import AggregationState
@@ -63,6 +62,7 @@ from metricflow.plan_conversion.spec_transforms import (
     SelectOnlyLinkableSpecs,
 )
 from metricflow.plan_conversion.sql_dataset import SqlDataSet
+from metricflow.plan_conversion.sql_join_builder import ColumnEqualityDescription, SqlQueryPlanJoinBuilder
 from metricflow.plan_conversion.time_spine import TimeSpineSource
 from metricflow.protocols.sql_client import SqlEngineAttributes, SqlEngine
 from metricflow.specs import (
@@ -111,73 +111,6 @@ logger = logging.getLogger(__name__)
 
 # The type of data set that present at a source node.
 SqlDataSetT = TypeVar("SqlDataSetT", bound=SqlDataSet)
-
-
-@dataclass(frozen=True)
-class ColumnEqualityDescription:
-    """Helper class to enumerate columns that should be equal between sources in a join."""
-
-    left_column_alias: str
-    right_column_alias: str
-
-
-def make_sql_join_description(
-    right_source_node: SqlSelectStatementNode,
-    left_source_alias: str,
-    right_source_alias: str,
-    column_equality_descriptions: Sequence[ColumnEqualityDescription],
-    join_type: SqlJoinType,
-    additional_on_conditions: Sequence[SqlExpressionNode] = tuple(),
-) -> SqlJoinDescription:
-    """Make a join description where the condition is a set of equality comparisons between columns.
-
-    Typically the columns in column_equality_descriptions are identifiers we are trying to match,
-    although they may include things like dimension partitions or time dimension columns where an
-    equality is expected.
-
-    Args:
-        right_source_node: node representing the join target, may be either a table or subquery
-        left_source_alias: string alias identifier for the join source
-        right_source_alias: string alias identifier for the join target
-        column_equality_descriptions: set of equality constraints for the ON statement
-        join_type: type of SQL join, e.g., LEFT, INNER, etc.
-        additional_on_conditions: set of additional constraints to add in the ON statement (via AND)
-    """
-    assert len(column_equality_descriptions) > 0
-
-    and_conditions: List[SqlExpressionNode] = []
-    for column_equality_description in column_equality_descriptions:
-        and_conditions.append(
-            SqlComparisonExpression(
-                left_expr=SqlColumnReferenceExpression(
-                    SqlColumnReference(
-                        table_alias=left_source_alias,
-                        column_name=column_equality_description.left_column_alias,
-                    )
-                ),
-                comparison=SqlComparison.EQUALS,
-                right_expr=SqlColumnReferenceExpression(
-                    SqlColumnReference(
-                        table_alias=right_source_alias,
-                        column_name=column_equality_description.right_column_alias,
-                    )
-                ),
-            )
-        )
-    and_conditions += additional_on_conditions
-
-    on_condition: SqlExpressionNode
-    if len(and_conditions) == 1:
-        on_condition = and_conditions[0]
-    else:
-        on_condition = SqlLogicalExpression(operator=SqlLogicalOperator.AND, args=tuple(and_conditions))
-
-    return SqlJoinDescription(
-        right_source=right_source_node,
-        right_source_alias=right_source_alias,
-        on_condition=on_condition,
-        join_type=join_type,
-    )
 
 
 def _make_validity_window_on_condition(
@@ -792,7 +725,7 @@ class DataflowToSqlQueryPlanConverter(Generic[SqlDataSetT], DataflowPlanNodeVisi
                 )
 
             sql_join_descs.append(
-                make_sql_join_description(
+                SqlQueryPlanJoinBuilder.make_sql_join_description(
                     right_source_node=right_data_set.sql_select_node,
                     left_source_alias=from_data_set_alias,
                     right_source_alias=right_data_set_alias,
@@ -1636,7 +1569,7 @@ class DataflowToSqlQueryPlanConverter(Generic[SqlDataSetT], DataflowPlanNodeVisi
         )
 
         join_data_set_alias = self._next_unique_table_alias()
-        sql_join_desc = make_sql_join_description(
+        sql_join_desc = SqlQueryPlanJoinBuilder.make_sql_join_description(
             right_source_node=row_filter_sql_select_node,
             left_source_alias=from_data_set_alias,
             right_source_alias=join_data_set_alias,
